requirements.txt
httpx
mcp
pydantic
pydantic-settings
python-dotenv
anyio
uvicorn 